help me

+++ src/train/train_function.py
@@ -88,7 +88,7 @@ def train_model(
     print(f"Starting training on {device} ")
     for batch_idx, (waveform, labels) in enumerate(dataloader):
         # Non-blocking transfer if pin_memory=True in dataloader
-        waveform = waveform.to(device, non_blocking=True)
+        waveform = waveform.to(device, non_blocking=True).float()
         labels = labels.to(device, non_blocking=True)
 
         # Determine split index
@@ -109,9 +109,9 @@ def train_model(
 
         # Apply spectrogram preprocessing if provided
         x_train = spectogram_model(
-            train_waveform).float() if spectogram_model else train_waveform.float()
-        x_val = spectogram_model(val_waveform).float() if (
-            spectogram_model and val_waveform is not None) else val_waveform.float()
+            train_waveform) if spectogram_model else train_waveform
+        x_val = spectogram_model(val_waveform) if (
+            spectogram_model and val_waveform is not None) else val_waveform
 
         model.train()
